+++ pragma-common/src/types.rs
@@ -1,4 +1,5 @@
 use serde::{Deserialize, Serialize};
+use std::ops::RangeInclusive;
 use utoipa::ToSchema;
 
 #[derive(Debug)]
@@ -73,5 +74,41 @@ impl Interval {
 
     pub fn to_seconds(&self) -> i64 {
         self.to_minutes() * 60
+    }
+}
+
+#[derive(Debug, Clone)]
+pub enum TimestampParam {
+    Single(u64),
+    Range(RangeInclusive<u64>),
+}
+
+impl From<u64> for TimestampParam {
+    fn from(ts: u64) -> Self {
+        TimestampParam::Single(ts)
+    }
+}
+
+pub mod timestamp_serde {
+    use super::*;
+    use serde::{self, Deserialize, Deserializer};
+
+    pub fn deserialize<'de, D>(deserializer: D) -> Result<Option<TimestampParam>, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        let s: Option<String> = Option::deserialize(deserializer)?;
+        if let Some(s) = s {
+            if let Some((start, end)) = s.split_once(',') {
+                let start = start.parse().map_err(serde::de::Error::custom)?;
+                let end = end.parse().map_err(serde::de::Error::custom)?;
+                Ok(Some(TimestampParam::Range(start..=end)))
+            } else {
+                let ts = s.parse().map_err(serde::de::Error::custom)?;
+                Ok(Some(TimestampParam::Single(ts)))
+            }
+        } else {
+            Ok(None)
+        }
     }
 }
